♻️ Move model types to each file
