+++ src/lib/server/db/schema.ts
@@ -1,7 +1,5 @@
 import { mysqlTable, int, varchar, datetime } from 'drizzle-orm/mysql-core';
 
-import type { InferSelectModel } from 'drizzle-orm';
-
 export const userTable = mysqlTable('users', {
 	id: int('id').primaryKey().autoincrement(),
 	googleId: varchar('google_id', { length: 255 }).notNull().unique(),
@@ -17,6 +15,3 @@ export const sessionTable = mysqlTable('sessions', {
 		.references(() => userTable.id),
 	expiresAt: datetime('expires_at').notNull()
 });
-
-export type User = InferSelectModel<typeof userTable>;
-export type Session = InferSelectModel<typeof sessionTable>;
